+++ src/main/java/kingdefense/backend/pieces/BlackKnight.java
@@ -4,18 +4,18 @@ import java.util.ArrayList;
 import java.util.Arrays;
 
 public class BlackKnight extends BlackPiece {
-    public BlackKnight(Integer x, Integer y, Integer health, Integer attack) {
+    public BlackKnight(Integer x, Integer y, Float health, Integer attack) {
         super(x, y, health, attack);
         this.priority = 1;
         this.attainableNb = 12;
         this.attainableX = new ArrayList<>(Arrays.asList(2, 2, -2, -2, 1, -1, 1, -1, 1, -1, 0, 0));
         this.attainableY = new ArrayList<>(Arrays.asList(1, -1, 1, -1, 2, 2, -2, -2, 0, 0, 1, -1));
     }
-    public BlackKnight(Integer health, Integer attack) {
+    public BlackKnight(Float health, Integer attack) {
         this(0, 0, health, attack);
     }
     public BlackKnight() {
-        this(0, 0, 3, 3);
+        this(0, 0, 3.f, 3);
     }
 
 	@Override
